Add files via upload
The code is for a scheduler of the generated topics from topic modelling (Natural language Processing) on a conference data set. It is an automatic scheduler which arranges all the generated topics and schedule events in consideration to the constraints. For details, please send queries to hoorain.malik@uq.net.au.